log the subscription ID on both the frontend and backend
