+++ frontend/controller/service-worker.js
@@ -8,7 +8,6 @@ import { HOURS_MILLIS } from '~/frontend/model/contracts/shared/time.js'
 import { GIMessage } from '~/shared/domains/chelonia/GIMessage.js'
 import { Secret } from '~/shared/domains/chelonia/Secret.js'
 import { deserializer, serializer } from '~/shared/serdes/index.js'
-import { cloneDeep } from '@model/contracts/shared/giLodash.js'
 
 const pwa = {
   deferredInstallPrompt: null,
@@ -203,7 +202,7 @@ sbp('sbp/selectors/register', {
 
       // Safari sometimes incorrectly reports 'prompt' when using `registration.pushManager.permissionState`
       let subscription = null
-      let subPrivVer = subscription
+      let endpoint = null
       if (notificationEnabled && granted) {
         try {
           subscription = await registration.pushManager.getSubscription()
@@ -212,12 +211,10 @@ sbp('sbp/selectors/register', {
             subscription = await registration.pushManager.subscribe(await sbp('push/getSubscriptionOptions'))
             newSub = true
           }
-          subPrivVer = cloneDeep(subscription)
           if (subscription?.endpoint) {
-            // hide the full endpoint from the logs for privacy
-            subPrivVer.endpoint = new URL(subPrivVer.endpoint).host
+            endpoint = new URL(subscription.endpoint).host // hide the full endpoint from the logs for privacy
           }
-          console.info(`[service-worker/setup-push-subscription] got ${newSub ? 'new' : 'existing'} subscription:`, subPrivVer)
+          console.info(`[service-worker/setup-push-subscription] got ${newSub ? 'new' : 'existing'} subscription:`, endpoint)
         } catch (e) {
           console.error('[service-worker/setup-push-subscription] error getting a subscription:', e)
           if (e?.message !== 'WebSocket connection is not open') {
@@ -237,7 +234,6 @@ sbp('sbp/selectors/register', {
       }
 
       try {
-        console.info('[service-worker/setup-push-subscription] reporting subscription:', subPrivVer)
         await sbp('push/reportExistingSubscription', subscription?.toJSON())
       } catch (e) {
         console.error('[service-worker/setup-push-subscription] error reporting subscription:', e)
